+++ CHANGELOG.md
@@ -15,12 +15,12 @@ Our versioning strategy is as follows:
 ### 🐛 Bug Fixes
 
 * Preview mode shows unpublishable content ([#410](https://github.com/Sitecore/content-sdk/pull/410))([416](https://github.com/Sitecore/content-sdk/pull/416))
+* `[core] [content]` Fix GraphQL client factory ignoring custom `fetch` and related options ([#418](https://github.com/Sitecore/content-sdk/pull/418))
 
 ## 2.0.1
 
 ### ✨ Bug fixes
 
-- `[core] [content]` Fix GraphQL client factory ignoring custom `fetch` and related options ([#418](https://github.com/Sitecore/content-sdk/pull/418))
 - Fix wrong versions and samples appearing when running `create-content-sdk-app` command.
   - `angular` sample no longer appears
   - `nextjs` samples use correct content-sdk dependencies for analytics when scaffolded